Remove lines which were not actually translated
As far as I can tell, I don't speak any of these languages and I'm merely making
educated guesses for them.
I've removed entries for which I'd reasonably expect there to be a better
translation. That does not mean that the most appropriate translation cannot be
identical to the original string, just that if I'd have to guess, I wouldn't
have guessed it.